+++ app/app.go
@@ -482,6 +482,7 @@ func NewEthermintApp(
 		tracer,
 		evmSs, nil,
 		allKeys,
+		nil,
 	)
 
 	// Create IBC Keeper