+++ tests/integration_tests/utils.py
@@ -40,6 +40,7 @@ TEST_CONTRACTS = {
     "TestMessageCall": "TestMessageCall.sol",
     "Calculator": "Calculator.sol",
     "Caller": "Caller.sol",
+    "Random": "Random.sol",
 }