+++ x/evm/keeper/state_transition.go
@@ -51,6 +51,7 @@ func (k *Keeper) NewEVM(
 	msg core.Message,
 	cfg *EVMConfig,
 	stateDB vm.StateDB,
+	random *common.Hash,
 ) *vm.EVM {
 	blockCtx := vm.BlockContext{
 		CanTransfer: core.CanTransfer,
@@ -62,7 +63,7 @@ func (k *Keeper) NewEVM(
 		Time:        uint64(ctx.BlockHeader().Time.Unix()),
 		Difficulty:  big.NewInt(0), // unused. Only required in PoW context
 		BaseFee:     cfg.BaseFee,
-		Random:      nil, // not supported
+		Random:      random, // not supported
 	}
 	if cfg.BlockOverrides != nil {
 		cfg.BlockOverrides.Apply(&blockCtx)
@@ -361,7 +362,12 @@ func (k *Keeper) ApplyMessageWithConfig(
 			return nil, errorsmod.Wrap(err, "failed to apply state override")
 		}
 	}
-	evm = k.NewEVM(ctx, msg, cfg, stateDB)
+	var random *common.Hash
+	if k.forkEnabledFunc(ctx) {
+		zero := common.BigToHash(big.NewInt(0))
+		random = &zero
+	}
+	evm = k.NewEVM(ctx, msg, cfg, stateDB, random)
 	leftoverGas := msg.GasLimit
 	sender := vm.AccountRef(msg.From)
 	// Allow the tracer captures the tx level events, mainly the gas consumption.